+++ src/jobs/DeleteProductItemsJob.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace kernpfad\commercedoofinder\jobs;
 
 use Craft;