+++ src/translations/en/commerce-doofinder.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'Syncing "{title}" to the Doofinder index' => 'Syncing "{title}" to the Doofinder index',
+    'Removing "{title}" from the Doofinder index' => 'Removing "{title}" from the Doofinder index',
+];